Extract prompt ingress reservation owner
